+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [#239] Extracted Dogu-CRD, Dogurestart-CRD and associated clients to own [repository](https://github.com/cloudogu/k8s-dogu-lib)
+- [#242] Prevent decreasing volume size
+- [#242] Update to dogu resource with new minimal volume size field
 
 ## [v3.5.1] - 2025-05-02